backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
langgraph==0.2.45
langchain-core==0.3.21
sqlmodel==0.0.22
python-dotenv==1.0.1
groq==0.13.1
python-dotenv==1.0.1

